Indev: Done Resetting NVMe Controller

+++ boot/kernel/gdt.h
@@ -12,11 +12,11 @@ typedef struct {
 typedef struct {
     uint16_t lim0 : 16;
     uint16_t base0 : 16;
-    uint8_t base1 : 8;
-    uint8_t a : 8;
-    uint8_t lim1 : 4;
-    uint8_t flag : 4;
-    uint8_t base2 : 8;
+    uint16_t base1 : 8;
+    uint16_t a : 8;
+    uint16_t lim1 : 4;
+    uint16_t flag : 4;
+    uint16_t base2 : 8;
 }__attribute__((packed)) SEGMENT_DESCRIPTOR;
 typedef struct {
     SEGMENT_DESCRIPTOR null;        // 0x00
